kubernetes/apps/eu-health-interop-lab/eu-gateway/requirements.txt
fastapi
uvicorn
requests
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
